updated the soln of 4th case

+++ data/solutions/P4/solution.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a, b;
-    scanf("%d %d", &a, &b);
-    printf("%d\n", a / b);
+    long long a, b;
+    scanf("%lld %lld", &a, &b);
+    printf("%lld\n", a / b);
     return 0;
 }
